requirements.txt
Flask==2.3.2
requests==2.31.0
pyaudio==0.2.14
wave==0.0.2
python-dotenv==1.0.0